Move TenantUUIDFromApiUrl from properties
Moved it to the only place where it should be used, which is the
migration/cleanup.
