+++ pkg/controllers/dynakube/dtpullsecret/generate.go
@@ -52,7 +52,7 @@ func (r *Reconciler) GenerateData() (map[string][]byte, error) {
 		return nil, errors.New("token secret does not contain a paas or api token, cannot generate docker config")
 	}
 
-	tenantUUID, err := r.dk.TenantUUIDFromConnectionInfoStatus()
+	tenantUUID, err := r.dk.TenantUUID()
 	if err != nil {
 		return nil, errors.WithMessage(err, "cannot generate docker config")
 	}
